src/requirements.txt
telebot
chromadb
nltk
requests
sentence-transformers
spacy
python-dotenv
python-docx
chardet